+++ app/Http/Controllers/PredicaController.php
@@ -10,26 +10,18 @@ use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Str;
 
 class PredicaController extends Controller
 {
-    function __construct()
-    {
-        //DOY LOS PERMISOS SEGUN EL ROL
-        $this->middleware('permission:ver-predicas|crear-predicas|editar-predicas|borrar-predicas', ['only' => ['index', 'cargarTabla']]);
-        $this->middleware(
-            'permission:ver-predicas',
-            ['only' => ['cargarTabla', 'obtener']]
-        );
-        $this->middleware('permission:crear-predicas', ['only' => ['guardar']]);
-        $this->middleware('permission:editar-predicas', ['only' => ['guardar']]);
-        $this->middleware('permission:borrar-predicas', ['only' => ['eliminar']]);
-    }
-
     public function index()
     {
         try {
-            return view('content.predicas.index');
+            if (Auth::user()->can('ver-predicas')) {
+                return view('content.predicas.index');
+            } else {
+                return view('content.error.403');
+            }
         } catch (Exception $e) {
 
             // return redirect()->route('content.roles.index')->with('error', true);
@@ -39,63 +31,72 @@ class PredicaController extends Controller
     public function cargarTabla()
     {
         try {
-            //OBTENGO LOS DATOS DEL USUARIO LOGUEADO Y SU ROL
-            $usuarioLogueado = User::with('roles')->find(Auth::user()->id);
-            $rolLogueado = $usuarioLogueado->roles->first();
+            if (Auth::user()->can('ver-predicas')) {
 
-            //HAGO EL SELECT A LA BASE DE DATOS PARA PODER MOSTRAR LOS REGISTROS
-            if ($usuarioLogueado->id != 1) { //no es superadmin
-                $predicas = Predica::where('id_usuario', $usuarioLogueado->id)->get();
-            } else {
-                $predicas = Predica::all();
-            }
-
-            $tablaHTML = '';
-
-            foreach ($predicas as $predica) {
-                if ($predica['estado'] === 0) {
-                    $predica['estado'] = '<span class="badge badge-warning">Inactivo</span>';
+                //HAGO EL SELECT A LA BASE DE DATOS PARA PODER MOSTRAR LOS REGISTROS
+                if (Auth::user()->id != 1) { //no es superadmin
+                    $predicas = Predica::where('id_usuario', Auth::user()->id)->get();
                 } else {
-                    $predica['estado'] = '<span class="badge badge-primary">Activo</span>';
+                    $predicas = Predica::all();
                 }
 
-                $btnEditar = '';
-                $btnEliminar = '';
+                $tablaHTML = '';
+                $contadorPagina = 1; // Inicializa un contador de página
+                foreach ($predicas as $predica) {
+                    if ($predica['estado'] === 0) {
+                        $predica['estado'] = '<span class="badge badge-warning">Inactivo</span>';
+                    } else {
+                        $predica['estado'] = '<span class="badge badge-primary">Activo</span>';
+                    }
+
+                    $btnVer = '';
+                    $btnEditar = '';
+                    $btnEliminar = '';
 
 
-                if ($rolLogueado->hasPermissionTo('editar-predicas')) {
-                    $btnEditar .= '<button type="button" onclick="editarPredicaModal(\'' . $predica['id'] . '\')" class="btn btn-icon btn-primary"><i class="fa fa-edit"></i></button>';
+                    if (Auth::user()->can('ver-predicas')) {
+                        $btnVer .= '<a href="' . url("/predicas/ver-predica?page={$contadorPagina}") . '" class="btn btn-icon btn-primary"><i class="fa fa-eye"></i></a>';
+                    }
+
+                    if (Auth::user()->can('editar-predicas')) {
+                        $btnEditar .= '<button type="button" onclick="editarPredicaModal(\'' . $predica['id'] . '\')" class="btn btn-icon btn-primary"><i class="fa fa-edit"></i></button>';
+                    }
+
+                    if (Auth::user()->can('borrar-predicas')) {
+                        $btnEliminar .= '<button type="button" onclick="eliminarPredicaModalConfirm(\'' . $predica['id'] . '\')" class="btn btn-icon btn-danger"><i class="fa fa-trash"></i></button>';
+                    }
+
+
+                    $tablaHTML .=  '<tr>' .
+                        '<td>' .
+                        $contadorPagina .
+                        '</td>' .
+                        '<td>' .
+                        $predica['nombre'] .
+                        '</td>' .
+                        '<td>' .
+                        $predica['descripcion_corta'] .
+                        '</td>' .
+                        '<td class="text-center">' .
+                        $predica['estado'] .
+                        '</td>' .
+                        '<td class="text-center">' .
+                        $btnVer . ' ' . $btnEditar . ' ' . $btnEliminar .
+                        '</td>' .
+                        '</tr>';
+
+                    $contadorPagina++; // Incrementa el contador de página
                 }
 
-                if ($rolLogueado->hasPermissionTo('borrar-predicas')) {
-                    $btnEliminar .= '<button type="button" onclick="eliminarPredicaModalConfirm(\'' . $predica['id'] . '\')" class="btn btn-icon btn-danger"><i class="fa fa-trash"></i></button>';
-                }
-
-
-                $tablaHTML .=  '<tr>' .
-                    '<td>' .
-                    $predica['nombre'] .
-                    '</td>' .
-                    '<td>' .
-                    $predica['descripcion_corta'] .
-                    '</td>' .
-                    '<td class="text-center">' .
-                    $predica['estado'] .
-                    '</td>' .
-                    '<td class="text-center">' .
-                    $btnEditar . ' ' . $btnEliminar .
-                    '</td>' .
-                    '</tr>';
+                return response()->json(
+                    [
+                        'estado' => true,
+                        'titulo' => 'Éxito',
+                        'msg' => 'Datos obtenidos correctamente',
+                        'datos' => $tablaHTML
+                    ]
+                );
             }
-
-            return response()->json(
-                [
-                    'estado' => true,
-                    'titulo' => 'Éxito',
-                    'msg' => 'Datos obtenidos correctamente',
-                    'datos' => $tablaHTML
-                ]
-            );
         } catch (Exception $e) {
             return response()->json(
                 [
@@ -108,90 +109,109 @@ class PredicaController extends Controller
         }
     }
 
+    public function verPredica()
+    {
+        try {
+            if (Auth::user()->can('ver-predicas')) {
+                $predicasPaginadas = Predica::where('estado', true)
+                    ->orderBy('id')
+                    ->paginate(1);
+
+                return view('content.predicas.ver', compact('predicasPaginadas'));
+            }
+        } catch (Exception $e) {
+            return view('content.error.404');
+        }
+    }
+
     public function guardar(Request $request)
     {
         $idPredica = $request->input('idPredica');
 
         try {
-            //SI EL ID ES 0, ENTONCES ES NUEVO REGISTRO
-            if (intval($idPredica) === 0) {
+            if (Auth::user()->can('crear-predicas', 'editar-predicas')) {
+                //SI EL ID ES 0, ENTONCES ES NUEVO REGISTRO
+                if (intval($idPredica) === 0) {
 
-                //VALIDO LOS CAMPOS
-                $validator = Validator::make($request->all(), [
-                    'nombre' => 'required|unique:predicas',
-                    'estado' => 'required',
-                    'descripcion_corta' => 'required',
-                    'descripcion_larga' => 'required',
-                ]);
-
-                //SI FALLA LA VALIDACION
-                if ($validator->fails()) {
-                    $errors = implode('<br>', $validator->errors()->all());
-
-                    return response()->json([
-                        'estado' => false,
-                        'titulo' => 'Error',
-                        'msg' => 'Revise los campos',
-                        'errors' => $errors
+                    //VALIDO LOS CAMPOS
+                    $validator = Validator::make($request->all(), [
+                        'nombre' => 'required|unique:predicas',
+                        'estado' => 'required',
+                        'descripcion_corta' => 'required',
+                        'descripcion_larga' => 'required',
                     ]);
-                }
 
-                //SI PASAN LAS VALIDACIONES, HAGO EL REGISTRO
-                $predica = Predica::create([
-                    'nombre' => $request->input('nombre'),
-                    'estado' => $request->input('estado'),
-                    'descripcion_corta' => $request->input('descripcion_corta'),
-                    'descripcion_larga' => $request->input('descripcion_larga'),
-                    'id_usuario' => Auth::user()->id,
-                ]);
+                    //SI FALLA LA VALIDACION
+                    if ($validator->fails()) {
+                        $errors = implode('<br>', $validator->errors()->all());
 
-                return response()->json(
-                    [
-                        'estado' => true,
-                        'titulo' => 'Éxito',
-                        'msg' => 'Registro realizado',
-                        'datos' => $predica
-                    ]
-                );
-            } else {
-                //ES EDICION
+                        return response()->json([
+                            'estado' => false,
+                            'titulo' => 'Error',
+                            'msg' => 'Revise los campos',
+                            'errors' => $errors
+                        ]);
+                    }
 
-                //VALIDO LOS CAMPOS
-                $validator = Validator::make($request->all(), [
-                    'nombre' => 'required|unique:predicas,nombre,' . $idPredica,
-                    'estado' => 'required',
-                    'descripcion_corta' => 'required',
-                    'descripcion_larga' => 'required',
-                ]);
-
-                //SI FALLA LA VALIDACION
-                if ($validator->fails()) {
-                    $errors = implode('<br>', $validator->errors()->all());
-
-                    return response()->json([
-                        'estado' => false,
-                        'titulo' => 'Error',
-                        'msg' => 'Revise los campos',
-                        'errors' => $errors
+                    //SI PASAN LAS VALIDACIONES, HAGO EL REGISTRO
+                    $predica = Predica::create([
+                        'nombre' => $request->input('nombre'),
+                        'slug' => Str::slug($request->input('nombre')),
+                        'estado' => $request->input('estado'),
+                        'descripcion_corta' => $request->input('descripcion_corta'),
+                        'descripcion_larga' => $request->input('descripcion_larga'),
+                        'id_usuario' => Auth::user()->id,
                     ]);
+
+                    return response()->json(
+                        [
+                            'estado' => true,
+                            'titulo' => 'Éxito',
+                            'msg' => 'Registro realizado',
+                            'datos' => $predica
+                        ]
+                    );
+                } else {
+                    //ES EDICION
+
+                    //VALIDO LOS CAMPOS
+                    $validator = Validator::make($request->all(), [
+                        'nombre' => 'required|unique:predicas,nombre,' . $idPredica,
+                        'estado' => 'required',
+                        'descripcion_corta' => 'required',
+                        'descripcion_larga' => 'required',
+                    ]);
+
+                    //SI FALLA LA VALIDACION
+                    if ($validator->fails()) {
+                        $errors = implode('<br>', $validator->errors()->all());
+
+                        return response()->json([
+                            'estado' => false,
+                            'titulo' => 'Error',
+                            'msg' => 'Revise los campos',
+                            'errors' => $errors
+                        ]);
+                    }
+
+                    $predica = Predica::find($idPredica);
+                    $predica->nombre = $request->input('nombre');
+                    $predica->slug = Str::slug($request->input('nombre'));
+                    $predica->estado = $request->input('estado');
+                    $predica->descripcion_corta = $request->input('descripcion_corta');
+                    $predica->descripcion_larga = $request->input('descripcion_larga');
+                    $predica->id_usuario = Auth::user()->id;
+                    $predica->save();
+
+                    return response()->json(
+                        [
+                            'estado' => true,
+                            'titulo' => 'Éxito',
+                            'msg' => 'Actualización realizada',
+                            'datos' => $predica
+                        ]
+                    );
                 }
-
-                $predica = Predica::find($idPredica);
-                $predica->nombre = $request->input('nombre');
-                $predica->estado = $request->input('estado');
-                $predica->descripcion_corta = $request->input('descripcion_corta');
-                $predica->descripcion_larga = $request->input('descripcion_larga');
-                $predica->id_usuario = Auth::user()->id;
-                $predica->save();
-
-                return response()->json(
-                    [
-                        'estado' => true,
-                        'titulo' => 'Éxito',
-                        'msg' => 'Actualización realizada',
-                        'datos' => $predica
-                    ]
-                );
             }
         } catch (Exception $e) {
             return response()->json(
@@ -208,17 +228,19 @@ class PredicaController extends Controller
     public function obtener($id)
     {
         try {
-            //OBTENGO LOS DATOS
-            $predica = Predica::find($id);
+            if (Auth::user()->can('ver-predicas')) {
+                //OBTENGO LOS DATOS
+                $predica = Predica::find($id);
 
-            return response()->json(
-                [
-                    'estado' => true,
-                    'titulo' => 'Éxito',
-                    'msg' => 'Datos obtenidos correctamente',
-                    'datos' => $predica
-                ]
-            );
+                return response()->json(
+                    [
+                        'estado' => true,
+                        'titulo' => 'Éxito',
+                        'msg' => 'Datos obtenidos correctamente',
+                        'datos' => $predica
+                    ]
+                );
+            }
         } catch (Exception $e) {
             return response()->json(
                 [
@@ -234,20 +256,22 @@ class PredicaController extends Controller
     public function eliminar(Request $request)
     {
         try {
-            //OBTENGO EL ID DEL POST
-            $idPredica = intval(json_decode($request->getContent(), true)['idPredica']);
+            if (Auth::user()->can('borrar-predicas')) {
+                //OBTENGO EL ID DEL POST
+                $idPredica = intval(json_decode($request->getContent(), true)['idPredica']);
 
-            //OBTENGO EL REGISTRO Y LO ELIMINO
-            Predica::find($idPredica)->delete();
+                //OBTENGO EL REGISTRO Y LO ELIMINO
+                Predica::find($idPredica)->delete();
 
-            return response()->json(
-                [
-                    'estado' => true,
-                    'titulo' => 'Éxito',
-                    'msg' => 'Eliminación realizada',
-                    'datos' => $idPredica
-                ]
-            );
+                return response()->json(
+                    [
+                        'estado' => true,
+                        'titulo' => 'Éxito',
+                        'msg' => 'Eliminación realizada',
+                        'datos' => $idPredica
+                    ]
+                );
+            }
         } catch (Exception $e) {
             return response()->json(
                 [
